Extracts `adhoc_tool` tests from `shell_command_test`
