+++ src/python/pants/backend/shell/util_rules/shell_command_test.py
@@ -4,17 +4,11 @@
 from __future__ import annotations
 
 import logging
-import os
 import shlex
 from textwrap import dedent
 
 import pytest
 
-from pants.backend.adhoc.adhoc_tool import GenerateFilesFromAdhocToolRequest
-from pants.backend.adhoc.adhoc_tool import rules as adhoc_tool_rules
-from pants.backend.adhoc.target_types import AdhocToolTarget
-from pants.backend.python.goals.run_python_source import rules as run_python_source_rules
-from pants.backend.python.target_types import PythonSourceTarget
 from pants.backend.shell.target_types import (
     ShellCommandRunTarget,
     ShellCommandTarget,
@@ -55,12 +49,9 @@ def rule_runner() -> RuleRunner:
         rules=[
             *archive.rules(),
             *shell_command_rules(),
-            *adhoc_tool_rules(),
             *source_files.rules(),
             *core_target_type_rules(),
-            *run_python_source_rules(),
             QueryRule(GeneratedSources, [GenerateFilesFromShellCommandRequest]),
-            QueryRule(GeneratedSources, [GenerateFilesFromAdhocToolRequest]),
             QueryRule(Process, [AdhocProcessRequest]),
             QueryRule(Process, [EnvironmentName, ShellCommandProcessFromTargetRequest]),
             QueryRule(RunRequest, [RunShellCommand]),
@@ -72,10 +63,8 @@ def rule_runner() -> RuleRunner:
             ShellCommandRunTarget,
             ShellCommandTestTarget,
             ShellSourcesGeneratorTarget,
-            AdhocToolTarget,
             ArchiveTarget,
             FilesGeneratorTarget,
-            PythonSourceTarget,
         ],
     )
     rule_runner.set_options([], env_inherit={"PATH"})
@@ -86,22 +75,14 @@ def assert_shell_command_result(
     rule_runner: RuleRunner,
     address: Address,
     expected_contents: dict[str, str],
-    generator_type: type[GenerateSourcesRequest] = GenerateFilesFromShellCommandRequest,
 ) -> None:
+    generator_type: type[GenerateSourcesRequest] = GenerateFilesFromShellCommandRequest
     target = rule_runner.get_target(address)
     result = rule_runner.request(GeneratedSources, [generator_type(EMPTY_SNAPSHOT, target)])
     assert result.snapshot.files == tuple(expected_contents)
     contents = rule_runner.request(DigestContents, [result.snapshot.digest])
     for fc in contents:
         assert fc.content == expected_contents[fc.path].encode()
-
-
-def assert_run_in_sandbox_result(
-    rule_runner: RuleRunner, address: Address, expected_contents: dict[str, str]
-) -> None:
-    return assert_shell_command_result(
-        rule_runner, address, expected_contents, GenerateFilesFromAdhocToolRequest
-    )
 
 
 def assert_logged(caplog, expect_logged=None):
@@ -724,115 +705,6 @@ def test_shell_command_extra_env_vars(caplog, rule_runner: RuleRunner) -> None:
     assert_logged(caplog, [(logging.INFO, "FOO=foo HELLO=world BAR=\n")])
 
 
-def test_run_runnable_in_sandbox(rule_runner: RuleRunner) -> None:
-    rule_runner.write_files(
-        {
-            "src/fruitcake.py": dedent(
-                """\
-                f = open("fruitcake.txt", "w")
-                f.write("fruitcake\\n")
-                f.close()
-                """
-            ),
-            "src/BUILD": dedent(
-                """\
-                python_source(
-                    source="fruitcake.py",
-                    name="fruitcake",
-                )
-
-                experimental_run_in_sandbox(
-                  name="run_fruitcake",
-                  runnable=":fruitcake",
-                  output_files=["fruitcake.txt"],
-                  root_output_directory=".",
-                )
-                """
-            ),
-        }
-    )
-
-    assert_run_in_sandbox_result(
-        rule_runner,
-        Address("src", target_name="run_fruitcake"),
-        expected_contents={"fruitcake.txt": "fruitcake\n"},
-    )
-
-
-def test_run_runnable_in_sandbox_with_workdir(rule_runner: RuleRunner) -> None:
-    rule_runner.write_files(
-        {
-            "src/fruitcake.py": dedent(
-                """\
-                f = open("src/fruitcake.txt", "w")
-                f.write("fruitcake\\n")
-                f.close()
-                """
-            ),
-            "src/BUILD": dedent(
-                """\
-                python_source(
-                    source="fruitcake.py",
-                    name="fruitcake",
-                )
-
-                experimental_run_in_sandbox(
-                  name="run_fruitcake",
-                  runnable=":fruitcake",
-                  output_files=["src/fruitcake.txt"],
-                  workdir="/",
-                )
-                """
-            ),
-        }
-    )
-
-    assert_run_in_sandbox_result(
-        rule_runner,
-        Address("src", target_name="run_fruitcake"),
-        expected_contents={"src/fruitcake.txt": "fruitcake\n"},
-    )
-
-
-def test_run_in_sandbox_capture_stdout_err(rule_runner: RuleRunner) -> None:
-    rule_runner.write_files(
-        {
-            "src/fruitcake.py": dedent(
-                """\
-                import sys
-                print("fruitcake")
-                print("inconceivable", file=sys.stderr)
-                """
-            ),
-            "src/BUILD": dedent(
-                """\
-                python_source(
-                    source="fruitcake.py",
-                    name="fruitcake",
-                )
-
-                experimental_run_in_sandbox(
-                  name="run_fruitcake",
-                  runnable=":fruitcake",
-                  stdout="stdout",
-                  stderr="stderr",
-                  root_output_directory=".",
-                )
-                """
-            ),
-        }
-    )
-
-    assert_run_in_sandbox_result(
-        rule_runner,
-        Address("src", target_name="run_fruitcake"),
-        expected_contents={
-            "stderr": "inconceivable\n",
-            "stdout": "fruitcake\n",
-        },
-    )
-
-
 def test_relative_directories(rule_runner: RuleRunner) -> None:
     rule_runner.write_files(
         {
@@ -901,54 +773,6 @@ def test_cannot_escape_build_root(rule_runner: RuleRunner) -> None:
             Address("src", target_name="quotes"),
             expected_contents={"../../invalid.txt": "foosh\n"},
         )
-
-
-@pytest.mark.parametrize(
-    ("workdir", "file_location"),
-    (
-        ("src", "src"),
-        (".", "src"),
-        ("./", "src"),
-        ("/", ""),
-        ("", ""),
-        ("/src", "src"),
-    ),
-)
-def test_working_directory_special_values(
-    rule_runner: RuleRunner, workdir: str, file_location: str
-) -> None:
-    rule_runner.write_files(
-        {
-            "src/fruitcake.py": dedent(
-                """\
-                f = open("fruitcake.txt", "w")
-                f.write("fruitcake\\n")
-                f.close()
-                """
-            ),
-            "src/BUILD": dedent(
-                f"""\
-                python_source(
-                    source="fruitcake.py",
-                    name="fruitcake",
-                )
-
-                experimental_run_in_sandbox(
-                  name="run_fruitcake",
-                  runnable=":fruitcake",
-                  output_files=["fruitcake.txt"],
-                  workdir="{workdir}",
-                )
-                """
-            ),
-        }
-    )
-
-    assert_run_in_sandbox_result(
-        rule_runner,
-        Address("src", target_name="run_fruitcake"),
-        expected_contents={os.path.join(file_location, "fruitcake.txt"): "fruitcake\n"},
-    )
 
 
 def test_missing_tool_called(
